GameOfLife

Documentation

type Cells = [Cell]

type Cell = (Coord, State)

type Coord = (Int, Int)

type Coords = [Coord]

data State

Constructors

Alive

Dead

type States = [State]

data Grid

Constructors

Grid Int Int Cells

nextGen

Arguments

:: Coords
Een lijst met levende coördinaten als oude generatie.

-> Coords
Een lijst met levende coördinaten als nieuwe generatie.

Deze functie berekent de nieuwe generatie van een grid op basis van een gegeven lijst met levende coördinaten.

nextGenCell

Arguments

:: Coord
De coördinaat waarvan de nieuwe generatie berekent moet worden.

-> Coords
Een lijst met cecoördinaten waarvan we weten dat ze levend zijn.

-> State
De nieuwe staat die de gegeven cel moet krijgen in de nieuwe generatie.

Deze functie berekent de nieuwe generatie van een coördinaat.

getNeighbours

Arguments

:: Coord
De coördinaat waarvan de buren berekent moeten worden.

-> Coords
De coördinaten van de buren van de gegeven coördinaat.

Deze functie berekent de coördinaten van de buren van een gegeven cel.

checkState

Arguments

:: Coord
De coördinaat waarvan je wilt weten welke staat hij heeft

-> Coords
De coördinaten waarvan je weet de ze levend zijn.

-> State
De berekende staat van de gegeven coördinaat.

Deze functie berekent of een coördinaat leven of dood is.

checkRules

Arguments

:: States
Een lijst met 9 staten waarvan de eerste staat in de lijst de oude staat is en de rest de buren zijn van de cell..

-> State
De nieuw berekende staat aan de hand van de buren van de cel

Deze functie berekent wat de nieuwe staat van een cel is aan de hand van een lijst met staten.

takeAliveCells

Arguments

:: Cells
Een lijst met cellen waarvan berekent word welke levend zijn.

-> Coords
Een lijst met coördinaten die de staat levend hebben.

Deze functie berekent de welke cellen levend zijn in een lijst van cellen.

buildGrid

Arguments

:: Int
Een getal die de grootte van het grid aangeeft in de hoogte.

-> Int
Een getal die de grootte van het grid aangeeft in de breedte.

-> Coords
Een lijst met coördinaten die levend gemaakt worden in het grid.

-> Grid
Het gecreeërde grid met de aangegeven grootte en welke cellen levend zijn.

Deze functie bouwt een grid waarbij je kan aangeven hoe groot hij moet zijn, en welke cellen levend moeten zijn.

setStatesAlive

Arguments

:: Coords
Een lijst met coördinaten die levend gemaakt moeten worden in het grid

-> Grid
Het gegeven grid waar cellen levend in gemaakt moeten worden.

-> Grid
Het grid waar de aangegeven cellen levend zijn gemaakt.

Deze functie maakt cellen in een grid levend aan de hand van een lijst met coördinaten.

setStateAlive

Arguments

:: Coord
De coördinaat die levend gemaakt moet worden in het gegeven grid.

-> Grid
Het gegeven grid waar de cel levend in gemaakt moet worden.

-> Grid
Het grid waar de aangegeven cel levend gemaakt is.

Deze functie maakt een aangegeven cell levend in een grid.

customMap

Arguments

:: (a -> b -> c)
De functie die gemapt word over de lijst.

-> [a]
De lijst waarover de functie gemapt word.

-> b
Het extra argument die nodig is voor de functie.

-> [c]
De lijst waarover gemapt is met de functie met het extra argument.

Dit is een geabstraheerde functie die bijna gelijk is aan de functie map. Het ver schil is dat je bij deze functie een extra argument kan meegeven aan de functie die je wilt mappen over de lijst.