Life

Description

Simulating Conway's Game of Life on a bounded board.

-- Get 15 generations from a 30x80 random world using 5 as seed:
gens = take 15 . iterate generation $ mkRandWorld (mkStdGen 5) 30 80
-- Print each generation:
mapM_ putStrLn . map showWorld $ gens

Types

data Cell

The state of a cell.

Constructors

Dead
cell is (still) dead

Alive
cell is (still) alive

Died
cell just died, i.e. was alive last generation

Born
cell was just born, i.e. was alive last generation

type World = Map Coord Cell

World grid

type Coord = (Height, Width)

Coordinates

type Height = Int

type Width = Int

Neighbourhood of a cell

neighbourhood :: Coord -> World -> World

Get the immediate neighbourhood of a cell (radius=1).

neighbours :: Coord -> Cell -> World -> Int

neighbours p life Count the number of cells within the neighbourhood whose state match life.

World generation

mkWorldFrom :: [Cell] -> Height -> Width -> World

mkWorldFrom v w h generates a w * h sized world grid with cell values from v. The values are repeated for all cells if the length of v is less than w * h.

mkBlankWorld :: Height -> Width -> World

Create a blank world with dead cells. Same as mkWorldFrom [Dead].

mkOverWorld :: Height -> Width -> World

Create an overpopulated world. Same as mkWorldFrom [Alive].

mkRandWorld :: StdGen -> Height -> Width -> World

Create a randomly populated world using the provided random generator.

Playing game of life

generation :: World -> World

generation w determines the next generation of a world w using the defaultRule.

generationWithRule :: Rule -> World -> World

generation rule world calculates the next generation of world, using rule.

Rules

type Rule = Cell -> Int -> Cell

Specifies a rule to determine the fate of a cell.

mkRule :: [Int] -> [Int] -> Rule

mkRule survive birth returns a new rule, where survive and birth specify cell counts for survival and birth of a cell:

mkRule [2,3] [3] == conwayRule

Common rules

defaultRule :: Rule

Default rule used by generation. Same as conwayRule.

conwayRule :: Rule

Conway's rule (23/3): A cell is born from 3 living cells. A cell only stays alive with 2 or 3 alive neighbours.

rule1357 :: Rule

Copy world (1357/1357)

Display

data LifeSign

String representation of cells.

Constructors

LifeSign

Fields

symDead :: String

symAlive :: String

symDied :: String

symBorn :: String

defaultLifeSign :: LifeSign

Default representation: Dead: " ", Alive: "●", Died: "◌", Born: "◍".

showWorld :: World -> String

Displays a grid of cells using defaultLifeSign.

showWorldUsing :: LifeSign -> World -> String

showWorldUsing w repr shows a grid of cells using a LifeSign repr as symbols for dead/alive cells.